+++ imports/main.te.js
@@ -1,7 +1,0 @@
-import assert from 'assert';
-describe('  - Inside', function(){
-  it("someting", function(){
-    console.log("Adasd");
-    assert.strictEqual(Meteor.isClient, false);
-  })
-})