+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## @2.0.02 - 2020-02-10
+## @2.0.30 - 2021-02-27
+- Added sass-variables integration
+- Tested with lighter vuetify imports using import Vuetify from 'vuetify/lib/framewor'. This reduces the initial client bundle size to 500Kb
+## @2.0.02 - 2021-02-10
 - Added HMR
 - Added SSR by default using vue-ssr
 - Added Vue-meta integrated with SSR for SEO